Fix casting float to int in getPrice() function in Ad entity

+++ src/Entity/Ad.php
@@ -50,7 +50,7 @@ class Ad
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
